requirements.txt
# Agentic Backend - LangChain Multi-Agent System
# New implementation using LangGraph for orchestration

# Create venv: python -m venv venv
# Activate: .\venv\Scripts\Activate.ps1
# Install: pip install -r requirements.txt

# ===== Core Framework =====
fastapi
uvicorn
python-multipart
pydantic

# ===== LangChain Stack (Latest) =====
langchain
langchain-core
langchain-groq
langgraph

# ===== LLM & Embeddings =====
groq
openai

# ===== Vector Store =====
pinecone

# ===== Database & Utils =====
python-dotenv
httpx>=0.27.0,<0.28.0

# ===== Data Processing =====
numpy
pandas

# ===== Image Processing (Receipt Scanning) =====
pillow

# ===== Optional: For Analytics =====
scikit-learn
